requirements.txt
pyobjc-framework-Speech
pyobjc-framework-AVFoundation
pyobjc-framework-Cocoa
pynput
sounddevice
numpy
